chore: split RichTextEditor into EditorToolbar, EditorRecording, EditorLinkDialog, EditorIcons, EditorStyles
