+++ src/components/editor/EditorStyles.css
@@ -1,0 +1,236 @@
+.is-editor-empty:first-child::before {
+  content: attr(data-placeholder);
+  float: left;
+  color: rgb(148 163 184);
+  pointer-events: none;
+  height: 0;
+}
+.ProseMirror {
+  min-height: 100%;
+}
+.ProseMirror:focus {
+  outline: none;
+}
+/* Remove focus outline and ring from TipTap wrapper */
+.tiptap:focus,
+.tiptap:focus-visible,
+.tiptap *:focus-visible {
+  outline: none;
+  box-shadow: none;
+  --tw-ring-shadow: 0 0 #0000;
+}
+
+/* Hide editor scrollbar unless hovered */
+.editor-scroll-container {
+  scrollbar-gutter: stable;
+}
+.editor-scroll-container::-webkit-scrollbar {
+  width: 4px;
+}
+.editor-scroll-container::-webkit-scrollbar-track {
+  background: transparent;
+}
+.editor-scroll-container::-webkit-scrollbar-thumb {
+  background: transparent;
+  border-radius: 9999px;
+  transition: background 0.2s;
+}
+.editor-scroll-container:hover::-webkit-scrollbar-thumb {
+  background: rgb(203 213 225);
+}
+.dark .editor-scroll-container:hover::-webkit-scrollbar-thumb {
+  background: rgb(71 85 105);
+}
+.ProseMirror p.is-editor-empty:first-child::before {
+  content: attr(data-placeholder);
+  float: left;
+  color: rgb(148 163 184);
+  pointer-events: none;
+  height: 0;
+}
+
+/* Paragraph spacing */
+.ProseMirror p {
+  margin: 0.75em 0;
+}
+
+/* Headings */
+.ProseMirror h2 {
+  font-size: 1.5em;
+  font-weight: 700;
+  line-height: 1.3;
+  margin: 1em 0 0.4em;
+  color: rgb(30 41 59);
+}
+.dark .ProseMirror h2 {
+  color: rgb(226 232 240);
+}
+.ProseMirror h3 {
+  font-size: 1.25em;
+  font-weight: 600;
+  line-height: 1.4;
+  margin: 0.8em 0 0.3em;
+  color: rgb(51 65 85);
+}
+.dark .ProseMirror h3 {
+  color: rgb(203 213 225);
+}
+
+/* Blockquote */
+.ProseMirror blockquote {
+  border-left: 3px solid rgb(139 92 246);
+  padding-left: 1em;
+  margin: 0.75em 0;
+  color: rgb(100 116 139);
+  font-style: italic;
+}
+.dark .ProseMirror blockquote {
+  border-left-color: rgb(167 139 250);
+  color: rgb(148 163 184);
+}
+
+/* Code block */
+.ProseMirror pre {
+  background: rgb(241 245 249);
+  border-radius: 0.5em;
+  padding: 0.75em 1em;
+  margin: 0.75em 0;
+  overflow-x: auto;
+  font-family: ui-monospace, SFMono-Regular, 'SF Mono', Menlo, Consolas, monospace;
+  font-size: 0.9em;
+  line-height: 1.5;
+  color: rgb(30 41 59);
+}
+.dark .ProseMirror pre {
+  background: rgb(30 41 59);
+  color: rgb(226 232 240);
+}
+.ProseMirror pre code {
+  background: none;
+  padding: 0;
+  border-radius: 0;
+  font-size: inherit;
+  color: inherit;
+}
+
+/* Inline code */
+.ProseMirror code {
+  background: rgb(241 245 249);
+  border-radius: 0.25em;
+  padding: 0.15em 0.35em;
+  font-family: ui-monospace, SFMono-Regular, 'SF Mono', Menlo, Consolas, monospace;
+  font-size: 0.9em;
+  color: rgb(139 92 246);
+}
+.dark .ProseMirror code {
+  background: rgb(30 41 59);
+  color: rgb(167 139 250);
+}
+
+/* Horizontal rule */
+.ProseMirror hr {
+  border: none;
+  border-top: 2px solid rgb(226 232 240);
+  margin: 1.5em 0;
+}
+.dark .ProseMirror hr {
+  border-top-color: rgb(51 65 85);
+}
+
+/* Lists */
+.ProseMirror ul {
+  list-style-type: disc;
+  padding-left: 1.5em;
+  margin: 0.5em 0;
+}
+.ProseMirror ol {
+  list-style-type: decimal;
+  padding-left: 1.5em;
+  margin: 0.5em 0;
+}
+.ProseMirror li {
+  margin: 0.25em 0;
+}
+.ProseMirror li p {
+  margin: 0;
+}
+
+/* Task list */
+.ProseMirror ul[data-type="taskList"] {
+  list-style: none;
+  padding-left: 0;
+  margin: 0.5em 0;
+}
+.ProseMirror ul[data-type="taskList"] li {
+  display: flex;
+  align-items: flex-start;
+  gap: 0.5em;
+  margin: 0.35em 0;
+}
+.ProseMirror ul[data-type="taskList"] li > label {
+  flex-shrink: 0;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin-top: 0.25em;
+  user-select: none;
+}
+.ProseMirror ul[data-type="taskList"] li > label input[type="checkbox"] {
+  appearance: none;
+  -webkit-appearance: none;
+  width: 1.15em;
+  height: 1.15em;
+  border: 2px solid rgb(203 213 225);
+  border-radius: 0.25em;
+  cursor: pointer;
+  position: relative;
+  transition: all 150ms ease;
+}
+.dark .ProseMirror ul[data-type="taskList"] li > label input[type="checkbox"] {
+  border-color: rgb(100 116 139);
+}
+.ProseMirror ul[data-type="taskList"] li > label input[type="checkbox"]:checked {
+  background: rgb(139 92 246);
+  border-color: rgb(139 92 246);
+}
+.dark .ProseMirror ul[data-type="taskList"] li > label input[type="checkbox"]:checked {
+  background: rgb(167 139 250);
+  border-color: rgb(167 139 250);
+}
+.ProseMirror ul[data-type="taskList"] li > label input[type="checkbox"]:checked::after {
+  content: '';
+  position: absolute;
+  top: 0.05em;
+  left: 0.25em;
+  width: 0.35em;
+  height: 0.6em;
+  border: solid white;
+  border-width: 0 2px 2px 0;
+  transform: rotate(45deg);
+}
+.ProseMirror ul[data-type="taskList"] li > div {
+  flex: 1;
+}
+.ProseMirror ul[data-type="taskList"] li[data-checked="true"] > div > p {
+  text-decoration: line-through;
+  color: rgb(148 163 184);
+}
+.dark .ProseMirror ul[data-type="taskList"] li[data-checked="true"] > div > p {
+  color: rgb(100 116 139);
+}
+
+/* Nested task lists */
+.ProseMirror ul[data-type="taskList"] ul[data-type="taskList"] {
+  margin: 0.25em 0;
+  padding-left: 1.5em;
+}
+
+/* Slash command menu — CSS variables for dark mode */
+:root {
+  --slash-menu-bg: white;
+  --slash-menu-border: rgb(226 232 240);
+}
+.dark {
+  --slash-menu-bg: rgb(15 23 42);
+  --slash-menu-border: rgb(51 65 85);
+}
