+++ src/components/editor/RichTextEditor.tsx
@@ -24,12 +24,16 @@ import TaskItem from '@tiptap/extension-task-item';
 import { FloatingToolbar } from './FloatingToolbar';
 import { EmojiPicker } from './EmojiPicker';
 import { SlashCommands } from './slashCommands';
-import { useSpeechToText, type STTState } from '../../hooks/useSpeechToText';
+import { useSpeechToText } from '../../hooks/useSpeechToText';
 import { useSettingsStore } from '../../stores/settingsStore';
 import { usePlatform } from '../../hooks/usePlatform';
 import { TranscriptPreviewOverlay } from '../transcript/TranscriptPreviewOverlay';
 import { MicrophonePermissionModal } from '../stt/MicrophonePermissionModal';
 import { MicrophoneBlockedModal } from '../stt/MicrophoneBlockedModal';
+import { CollapsibleToolbar } from './EditorToolbar';
+import { RecordingStrip } from './EditorRecording';
+import { LinkDialog } from './EditorLinkDialog';
+import './EditorStyles.css';
 
 interface RichTextEditorProps {
   value: string;
@@ -488,937 +492,6 @@ export function RichTextEditor({
         isOpen={permissionModal === 'blocked'}
         onDismiss={dismissPermissionModal}
       />
-
-      {/* Editor styling — restore list styles stripped by Tailwind preflight */}
-      <style>{`
-        .is-editor-empty:first-child::before {
-          content: attr(data-placeholder);
-          float: left;
-          color: rgb(148 163 184);
-          pointer-events: none;
-          height: 0;
-        }
-        .ProseMirror {
-          min-height: 100%;
-        }
-        .ProseMirror:focus {
-          outline: none;
-        }
-        /* Remove focus outline and ring from TipTap wrapper */
-        .tiptap:focus,
-        .tiptap:focus-visible,
-        .tiptap *:focus-visible {
-          outline: none;
-          box-shadow: none;
-          --tw-ring-shadow: 0 0 #0000;
-        }
-
-        /* Hide editor scrollbar unless hovered */
-        .editor-scroll-container {
-          scrollbar-gutter: stable;
-        }
-        .editor-scroll-container::-webkit-scrollbar {
-          width: 4px;
-        }
-        .editor-scroll-container::-webkit-scrollbar-track {
-          background: transparent;
-        }
-        .editor-scroll-container::-webkit-scrollbar-thumb {
-          background: transparent;
-          border-radius: 9999px;
-          transition: background 0.2s;
-        }
-        .editor-scroll-container:hover::-webkit-scrollbar-thumb {
-          background: rgb(203 213 225);
-        }
-        .dark .editor-scroll-container:hover::-webkit-scrollbar-thumb {
-          background: rgb(71 85 105);
-        }
-        .ProseMirror p.is-editor-empty:first-child::before {
-          content: attr(data-placeholder);
-          float: left;
-          color: rgb(148 163 184);
-          pointer-events: none;
-          height: 0;
-        }
-
-        /* Paragraph spacing */
-        .ProseMirror p {
-          margin: 0.75em 0;
-        }
-
-        /* Headings */
-        .ProseMirror h2 {
-          font-size: 1.5em;
-          font-weight: 700;
-          line-height: 1.3;
-          margin: 1em 0 0.4em;
-          color: rgb(30 41 59);
-        }
-        .dark .ProseMirror h2 {
-          color: rgb(226 232 240);
-        }
-        .ProseMirror h3 {
-          font-size: 1.25em;
-          font-weight: 600;
-          line-height: 1.4;
-          margin: 0.8em 0 0.3em;
-          color: rgb(51 65 85);
-        }
-        .dark .ProseMirror h3 {
-          color: rgb(203 213 225);
-        }
-
-        /* Blockquote */
-        .ProseMirror blockquote {
-          border-left: 3px solid rgb(139 92 246);
-          padding-left: 1em;
-          margin: 0.75em 0;
-          color: rgb(100 116 139);
-          font-style: italic;
-        }
-        .dark .ProseMirror blockquote {
-          border-left-color: rgb(167 139 250);
-          color: rgb(148 163 184);
-        }
-
-        /* Code block */
-        .ProseMirror pre {
-          background: rgb(241 245 249);
-          border-radius: 0.5em;
-          padding: 0.75em 1em;
-          margin: 0.75em 0;
-          overflow-x: auto;
-          font-family: ui-monospace, SFMono-Regular, 'SF Mono', Menlo, Consolas, monospace;
-          font-size: 0.9em;
-          line-height: 1.5;
-          color: rgb(30 41 59);
-        }
-        .dark .ProseMirror pre {
-          background: rgb(30 41 59);
-          color: rgb(226 232 240);
-        }
-        .ProseMirror pre code {
-          background: none;
-          padding: 0;
-          border-radius: 0;
-          font-size: inherit;
-          color: inherit;
-        }
-
-        /* Inline code */
-        .ProseMirror code {
-          background: rgb(241 245 249);
-          border-radius: 0.25em;
-          padding: 0.15em 0.35em;
-          font-family: ui-monospace, SFMono-Regular, 'SF Mono', Menlo, Consolas, monospace;
-          font-size: 0.9em;
-          color: rgb(139 92 246);
-        }
-        .dark .ProseMirror code {
-          background: rgb(30 41 59);
-          color: rgb(167 139 250);
-        }
-
-        /* Horizontal rule */
-        .ProseMirror hr {
-          border: none;
-          border-top: 2px solid rgb(226 232 240);
-          margin: 1.5em 0;
-        }
-        .dark .ProseMirror hr {
-          border-top-color: rgb(51 65 85);
-        }
-
-        /* Lists */
-        .ProseMirror ul {
-          list-style-type: disc;
-          padding-left: 1.5em;
-          margin: 0.5em 0;
-        }
-        .ProseMirror ol {
-          list-style-type: decimal;
-          padding-left: 1.5em;
-          margin: 0.5em 0;
-        }
-        .ProseMirror li {
-          margin: 0.25em 0;
-        }
-        .ProseMirror li p {
-          margin: 0;
-        }
-
-        /* Task list */
-        .ProseMirror ul[data-type="taskList"] {
-          list-style: none;
-          padding-left: 0;
-          margin: 0.5em 0;
-        }
-        .ProseMirror ul[data-type="taskList"] li {
-          display: flex;
-          align-items: flex-start;
-          gap: 0.5em;
-          margin: 0.35em 0;
-        }
-        .ProseMirror ul[data-type="taskList"] li > label {
-          flex-shrink: 0;
-          display: flex;
-          align-items: center;
-          justify-content: center;
-          margin-top: 0.25em;
-          user-select: none;
-        }
-        .ProseMirror ul[data-type="taskList"] li > label input[type="checkbox"] {
-          appearance: none;
-          -webkit-appearance: none;
-          width: 1.15em;
-          height: 1.15em;
-          border: 2px solid rgb(203 213 225);
-          border-radius: 0.25em;
-          cursor: pointer;
-          position: relative;
-          transition: all 150ms ease;
-        }
-        .dark .ProseMirror ul[data-type="taskList"] li > label input[type="checkbox"] {
-          border-color: rgb(100 116 139);
-        }
-        .ProseMirror ul[data-type="taskList"] li > label input[type="checkbox"]:checked {
-          background: rgb(139 92 246);
-          border-color: rgb(139 92 246);
-        }
-        .dark .ProseMirror ul[data-type="taskList"] li > label input[type="checkbox"]:checked {
-          background: rgb(167 139 250);
-          border-color: rgb(167 139 250);
-        }
-        .ProseMirror ul[data-type="taskList"] li > label input[type="checkbox"]:checked::after {
-          content: '';
-          position: absolute;
-          top: 0.05em;
-          left: 0.25em;
-          width: 0.35em;
-          height: 0.6em;
-          border: solid white;
-          border-width: 0 2px 2px 0;
-          transform: rotate(45deg);
-        }
-        .ProseMirror ul[data-type="taskList"] li > div {
-          flex: 1;
-        }
-        .ProseMirror ul[data-type="taskList"] li[data-checked="true"] > div > p {
-          text-decoration: line-through;
-          color: rgb(148 163 184);
-        }
-        .dark .ProseMirror ul[data-type="taskList"] li[data-checked="true"] > div > p {
-          color: rgb(100 116 139);
-        }
-
-        /* Nested task lists */
-        .ProseMirror ul[data-type="taskList"] ul[data-type="taskList"] {
-          margin: 0.25em 0;
-          padding-left: 1.5em;
-        }
-
-        /* Slash command menu — CSS variables for dark mode */
-        :root {
-          --slash-menu-bg: white;
-          --slash-menu-border: rgb(226 232 240);
-        }
-        .dark {
-          --slash-menu-bg: rgb(15 23 42);
-          --slash-menu-border: rgb(51 65 85);
-        }
-      `}</style>
     </div>
-  );
-}
-
-// ============================================
-// Recording Indicator Strip
-// ============================================
-
-function formatElapsed(seconds: number): string {
-  const m = Math.floor(seconds / 60);
-  const s = seconds % 60;
-  return `${String(m).padStart(2, '0')}:${String(s).padStart(2, '0')}`;
-}
-
-const prefersReducedMotion =
-  typeof window !== 'undefined' &&
-  window.matchMedia('(prefers-reduced-motion: reduce)').matches;
-
-function WaveformBars({ active }: { active: boolean }) {
-  const bars = Array.from({ length: 20 }, (_, i) => i);
-
-  if (!active || prefersReducedMotion) {
-    return (
-      <div className="flex items-center gap-px h-5" aria-hidden="true">
-        {bars.map((i) => (
-          <div
-            key={i}
-            className="w-px rounded-full bg-red-400 dark:bg-red-500"
-            style={{ height: '6px' }}
-          />
-        ))}
-      </div>
-    );
-  }
-
-  return (
-    <div className="flex items-center gap-px h-5" aria-hidden="true">
-      {bars.map((i) => (
-        <div
-          key={i}
-          className="w-px rounded-full bg-red-400 dark:bg-red-500 animate-waveform"
-          style={{
-            height: '6px',
-            animationDelay: `${(i * 80) % 600}ms`,
-            animationDuration: `${600 + (i * 37) % 400}ms`,
-          }}
-        />
-      ))}
-    </div>
-  );
-}
-
-interface RecordingStripProps {
-  state: STTState;
-  elapsedSeconds: number;
-  onStop: () => void;
-  onCancel: () => void;
-}
-
-function RecordingStrip({ state, elapsedSeconds, onStop, onCancel }: RecordingStripProps) {
-  const isRecording = state === 'recording';
-  const isTranscribing = state === 'transcribing' || state === 'processing' || state === 'formatting';
-
-  if (!isRecording && !isTranscribing) return null;
-
-  return (
-    <div className="flex-shrink-0 flex items-center gap-3 px-3 border-b border-red-100 dark:border-red-900/30 bg-red-50/60 dark:bg-red-950/20 h-10">
-      {/* Waveform */}
-      {isRecording ? (
-        <WaveformBars active />
-      ) : (
-        <svg className="w-4 h-4 animate-spin text-violet-500" fill="none" viewBox="0 0 24 24">
-          <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4" />
-          <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z" />
-        </svg>
-      )}
-
-      {/* Status text + timer */}
-      <span className="text-xs text-red-600 dark:text-red-400 tabular-nums flex-1">
-        {isRecording ? (
-          <>{formatElapsed(elapsedSeconds)} &mdash; Recording&hellip;</>
-        ) : (
-          'Transcribing\u2026'
-        )}
-      </span>
-
-      {/* Stop / Cancel buttons */}
-      {isRecording && (
-        <>
-          <button
-            type="button"
-            onClick={onStop}
-            className="text-xs font-medium px-2 py-0.5 rounded bg-red-500 hover:bg-red-600 text-white transition-colors duration-150"
-            aria-label="Stop recording"
-          >
-            Stop
-          </button>
-          <button
-            type="button"
-            onClick={onCancel}
-            className="text-xs text-slate-400 hover:text-slate-600 dark:text-slate-500 dark:hover:text-slate-300 transition-colors duration-150"
-            aria-label="Cancel recording"
-          >
-            Cancel
-          </button>
-        </>
-      )}
-    </div>
-  );
-}
-
-// Collapsible Toolbar
-// ============================================
-
-interface CollapsibleToolbarProps {
-  editor: Editor;
-  onFormat: (command: string, value?: string) => void;
-  getFormatState: () => Record<string, boolean>;
-  expanded: boolean;
-  onToggle: (expanded: boolean) => void;
-  onLinkClick: () => void;
-  onEmojiClick: () => void;
-  // Speech-to-text
-  sttReady?: boolean; // true if enabled AND model downloaded
-  sttState?: STTState;
-  sttError?: string | null;
-  onMicClick?: () => void;
-  onMicCancel?: () => void;
-  quickCapture?: boolean;
-  onToggleQuickCapture?: () => void;
-}
-
-function CollapsibleToolbar({
-  editor,
-  onFormat,
-  getFormatState,
-  expanded,
-  onToggle,
-  onLinkClick,
-  onEmojiClick,
-  sttReady = false,
-  sttState = 'idle',
-  sttError,
-  onMicClick,
-  onMicCancel,
-  quickCapture = false,
-  onToggleQuickCapture,
-}: CollapsibleToolbarProps) {
-  const [formatState, setFormatState] = useState<Record<string, boolean>>({});
-
-  // Track format state on selection/transaction changes
-  useEffect(() => {
-    const update = () => setFormatState(getFormatState());
-    editor.on('selectionUpdate', update);
-    editor.on('transaction', update);
-    return () => {
-      editor.off('selectionUpdate', update);
-      editor.off('transaction', update);
-    };
-  }, [editor, getFormatState]);
-
-  return (
-    <div className="flex-shrink-0 border-b border-slate-100 dark:border-slate-800">
-      {/* Toggle bar */}
-      <button
-        type="button"
-        onClick={() => onToggle(!expanded)}
-        className="w-full flex items-center gap-1.5 px-2 py-1.5 text-xs text-slate-400 dark:text-slate-500 hover:text-slate-500 dark:hover:text-slate-400 transition-colors"
-      >
-        <svg
-          className={`w-3 h-3 transition-transform duration-200 ${expanded ? 'rotate-180' : ''}`}
-          fill="none"
-          viewBox="0 0 24 24"
-          stroke="currentColor"
-          strokeWidth={2}
-        >
-          <path strokeLinecap="round" strokeLinejoin="round" d="M19 9l-7 7-7-7" />
-        </svg>
-        <span>Formatting</span>
-      </button>
-
-      {/* Expandable button row — overflow-x-auto so mobile can scroll horizontally */}
-      <div
-        className={`overflow-y-hidden transition-all duration-300 ease-out ${expanded ? 'max-h-16 opacity-100' : 'max-h-0 opacity-0'}`}
-      >
-        <div className="overflow-x-auto scrollbar-hide">
-        <div className="flex items-center gap-0.5 px-2 pb-2 flex-nowrap">
-          {/* Text formatting */}
-          <ToolbarBtn
-            icon={<TBBoldIcon />}
-            label="Bold (Ctrl+B)"
-            isActive={formatState.bold}
-            onClick={() => onFormat('bold')}
-          />
-          <ToolbarBtn
-            icon={<TBItalicIcon />}
-            label="Italic (Ctrl+I)"
-            isActive={formatState.italic}
-            onClick={() => onFormat('italic')}
-          />
-          <ToolbarBtn
-            icon={<TBUnderlineIcon />}
-            label="Underline (Ctrl+U)"
-            isActive={formatState.underline}
-            onClick={() => onFormat('underline')}
-          />
-          <ToolbarBtn
-            icon={<TBStrikeIcon />}
-            label="Strikethrough"
-            isActive={formatState.strike}
-            onClick={() => onFormat('strike')}
-          />
-
-          <TBDivider />
-
-          {/* Lists */}
-          <ToolbarBtn
-            icon={<TBBulletListIcon />}
-            label="Bullet list"
-            isActive={formatState.bulletList}
-            onClick={() => onFormat('bulletList')}
-          />
-          <ToolbarBtn
-            icon={<TBOrderedListIcon />}
-            label="Numbered list"
-            isActive={formatState.orderedList}
-            onClick={() => onFormat('orderedList')}
-          />
-
-          <TBDivider />
-
-          {/* Blocks */}
-          <ToolbarBtn
-            icon={<TBHeading2Icon />}
-            label="Heading 2"
-            isActive={formatState.heading2}
-            onClick={() => onFormat('heading2')}
-          />
-          <ToolbarBtn
-            icon={<TBHeading3Icon />}
-            label="Heading 3"
-            isActive={formatState.heading3}
-            onClick={() => onFormat('heading3')}
-          />
-          <ToolbarBtn
-            icon={<TBBlockquoteIcon />}
-            label="Blockquote"
-            isActive={formatState.blockquote}
-            onClick={() => onFormat('blockquote')}
-          />
-          <ToolbarBtn
-            icon={<TBTaskListIcon />}
-            label="Task list"
-            isActive={formatState.taskList}
-            onClick={() => onFormat('taskList')}
-          />
-
-          <TBDivider />
-
-          {/* Insert */}
-          <ToolbarBtn
-            icon={<TBLinkIcon />}
-            label="Link (Ctrl+K)"
-            isActive={formatState.link}
-            onClick={onLinkClick}
-          />
-          <ToolbarBtn
-            icon={<TBEmojiIcon />}
-            label="Emoji"
-            onClick={onEmojiClick}
-          />
-
-          {/* Speech-to-text: quick capture toggle + mic button */}
-          {onMicClick && (
-            <>
-              <TBDivider />
-              {onToggleQuickCapture && (
-                <QuickCaptureToggle
-                  active={quickCapture}
-                  onToggle={onToggleQuickCapture}
-                />
-              )}
-              <MicButton
-                state={sttState}
-                error={sttError}
-                onClick={onMicClick}
-                onCancel={onMicCancel}
-                isReady={sttReady}
-              />
-            </>
-          )}
-        </div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function ToolbarBtn({
-  icon,
-  label,
-  isActive,
-  onClick,
-}: {
-  icon: React.ReactNode;
-  label: string;
-  isActive?: boolean;
-  onClick: () => void;
-}) {
-  return (
-    <button
-      type="button"
-      onMouseDown={(e) => e.preventDefault()}
-      onClick={(e) => {
-        e.preventDefault();
-        onClick();
-      }}
-      className={`
-        p-1.5 rounded transition-all duration-150 active:scale-90
-        ${isActive
-          ? 'bg-violet-100 dark:bg-violet-900/30 text-violet-600 dark:text-violet-400'
-          : 'text-slate-400 dark:text-slate-500 hover:bg-slate-100 dark:hover:bg-slate-800 hover:text-slate-600 dark:hover:text-slate-300'
-        }
-      `}
-      title={label}
-    >
-      {icon}
-    </button>
-  );
-}
-
-function TBDivider() {
-  return <div className="w-px h-5 bg-slate-200 dark:bg-slate-700 mx-1" />;
-}
-
-// Mic button for speech-to-text
-function MicButton({
-  state,
-  error,
-  onClick,
-  onCancel,
-  isReady = false,
-}: {
-  state: STTState;
-  error?: string | null;
-  onClick: () => void;
-  onCancel?: () => void;
-  isReady?: boolean;
-}) {
-  const isRecording = state === 'recording';
-  const isFormatting = state === 'formatting';
-  const isProcessing = state === 'processing' || state === 'transcribing' || state === 'requesting' || isFormatting;
-
-  // Determine button appearance based on state
-  let buttonClass = 'text-slate-400 dark:text-slate-500 hover:bg-slate-100 dark:hover:bg-slate-800 hover:text-slate-600 dark:hover:text-slate-300';
-  let title = isReady ? 'Start dictation' : 'Set up speech-to-text';
-
-  if (!isReady) {
-    // Not configured - show setup hint
-    buttonClass = 'text-slate-300 dark:text-slate-600 hover:bg-violet-50 dark:hover:bg-violet-900/20 hover:text-violet-500 dark:hover:text-violet-400';
-  } else if (isRecording) {
-    buttonClass = 'bg-red-100 dark:bg-red-900/30 text-red-600 dark:text-red-400 animate-pulse';
-    title = 'Stop recording';
-  } else if (isFormatting) {
-    buttonClass = 'bg-amber-100 dark:bg-amber-900/30 text-amber-500 dark:text-amber-400';
-    title = 'Formatting…';
-  } else if (isProcessing) {
-    buttonClass = 'bg-violet-100 dark:bg-violet-900/30 text-violet-600 dark:text-violet-400';
-    title = state === 'transcribing' ? 'Transcribing...' : 'Processing...';
-  } else if (error) {
-    buttonClass = 'bg-amber-100 dark:bg-amber-900/30 text-amber-600 dark:text-amber-400';
-    title = error;
-  }
-
-  return (
-    <div className="relative">
-      <button
-        type="button"
-        onMouseDown={(e) => e.preventDefault()}
-        onClick={(e) => {
-          e.preventDefault();
-          onClick();
-        }}
-        disabled={isProcessing}
-        className={`p-1.5 rounded transition-all duration-150 active:scale-90 ${buttonClass} ${isProcessing ? 'cursor-wait' : ''}`}
-        title={title}
-      >
-        {isProcessing ? (
-          <span className="flex items-center gap-1">
-            <svg className={`w-4 h-4 animate-spin ${isFormatting ? 'text-amber-500' : ''}`} fill="none" viewBox="0 0 24 24">
-              <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4" />
-              <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z" />
-            </svg>
-            {isFormatting && <span className="text-xs text-amber-500">Formatting…</span>}
-          </span>
-        ) : (
-          <TBMicIcon isRecording={isRecording} />
-        )}
-      </button>
-
-      {/* Cancel button shown during recording */}
-      {isRecording && onCancel && (
-        <button
-          type="button"
-          onClick={(e) => {
-            e.preventDefault();
-            e.stopPropagation();
-            onCancel();
-          }}
-          className="absolute -top-1 -right-1 w-4 h-4 rounded-full bg-slate-200 dark:bg-slate-700 text-slate-500 dark:text-slate-400 hover:bg-slate-300 dark:hover:bg-slate-600 flex items-center justify-center text-xs"
-          title="Cancel recording"
-        >
-          <svg className="w-2.5 h-2.5" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={3}>
-            <path strokeLinecap="round" strokeLinejoin="round" d="M6 18L18 6M6 6l12 12" />
-          </svg>
-        </button>
-      )}
-    </div>
-  );
-}
-
-function TBMicIcon({ isRecording }: { isRecording?: boolean }) {
-  return (
-    <svg className="w-4 h-4" fill={isRecording ? 'currentColor' : 'none'} viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-      <path strokeLinecap="round" strokeLinejoin="round" d="M12 18.75a6 6 0 006-6v-1.5m-6 7.5a6 6 0 01-6-6v-1.5m6 7.5v3.75m-3.75 0h7.5M12 15.75a3 3 0 01-3-3V4.5a3 3 0 116 0v8.25a3 3 0 01-3 3z" />
-    </svg>
-  );
-}
-
-/**
- * Quick capture toggle — bolt/lightning icon. When active, transcription
- * bypasses all formatting layers and inserts raw text immediately.
- */
-function QuickCaptureToggle({
-  active,
-  onToggle,
-}: {
-  active: boolean;
-  onToggle: () => void;
-}) {
-  return (
-    <button
-      type="button"
-      onMouseDown={(e) => e.preventDefault()}
-      onClick={(e) => {
-        e.preventDefault();
-        onToggle();
-      }}
-      title="Quick capture (bypass formatting)"
-      aria-pressed={active}
-      className={[
-        'p-1.5 rounded transition-all duration-150 active:scale-90 min-w-[44px] min-h-[44px] flex items-center justify-center',
-        active
-          ? 'text-amber-500 bg-amber-50 dark:bg-amber-900/20'
-          : 'text-gray-400 hover:text-gray-600 dark:hover:text-gray-300 hover:bg-slate-100 dark:hover:bg-slate-800',
-      ].join(' ')}
-    >
-      {/* Lightning bolt / bolt icon */}
-      <svg className="w-4 h-4" fill={active ? 'currentColor' : 'none'} viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-        <path strokeLinecap="round" strokeLinejoin="round" d="M13 10V3L4 14h7v7l9-11h-7z" />
-      </svg>
-    </button>
-  );
-}
-
-// ============================================
-// Link Dialog
-// ============================================
-
-interface LinkDialogProps {
-  initialUrl: string;
-  onSubmit: (url: string) => void;
-  onClose: () => void;
-}
-
-function LinkDialog({ initialUrl, onSubmit, onClose }: LinkDialogProps) {
-  const [url, setUrl] = useState(initialUrl || 'https://');
-  const inputRef = useRef<HTMLInputElement>(null);
-
-  // Auto-focus and select URL on open
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      inputRef.current?.focus();
-      inputRef.current?.select();
-    }, 50);
-    return () => clearTimeout(timer);
-  }, []);
-
-  // Close on Escape
-  useEffect(() => {
-    const handleKeyDown = (e: KeyboardEvent) => {
-      if (e.key === 'Escape') onClose();
-    };
-    document.addEventListener('keydown', handleKeyDown);
-    return () => document.removeEventListener('keydown', handleKeyDown);
-  }, [onClose]);
-
-  const handleSubmit = (e: React.FormEvent) => {
-    e.preventDefault();
-    const trimmed = url.trim();
-    onSubmit(trimmed === 'https://' ? '' : trimmed);
-  };
-
-  const handleRemoveLink = () => {
-    onSubmit('');
-  };
-
-  return (
-    <div
-      className="fixed inset-0 z-50 flex items-center justify-center p-4 bg-black/50 backdrop-blur-sm"
-      onClick={onClose}
-    >
-      <div
-        className="w-full max-w-sm bg-white dark:bg-slate-900 rounded-2xl shadow-2xl animate-slide-up"
-        onClick={(e) => e.stopPropagation()}
-      >
-        {/* Header */}
-        <div className="flex items-center justify-between px-5 pt-5 pb-3">
-          <div className="flex items-center gap-2.5">
-            <div className="w-8 h-8 rounded-lg bg-violet-100 dark:bg-violet-900/30 flex items-center justify-center">
-              <svg className="w-4 h-4 text-violet-600 dark:text-violet-400" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-                <path strokeLinecap="round" strokeLinejoin="round" d="M13.828 10.172a4 4 0 00-5.656 0l-4 4a4 4 0 105.656 5.656l1.102-1.101m-.758-4.899a4 4 0 005.656 0l4-4a4 4 0 00-5.656-5.656l-1.1 1.1" />
-              </svg>
-            </div>
-            <h2 className="text-base font-semibold text-slate-800 dark:text-slate-100">
-              {initialUrl ? 'Edit Link' : 'Add Link'}
-            </h2>
-          </div>
-          <button
-            type="button"
-            onClick={onClose}
-            className="p-1.5 rounded-lg text-slate-400 dark:text-slate-500 hover:bg-slate-100 dark:hover:bg-slate-800 transition-colors"
-            aria-label="Close"
-          >
-            <svg className="w-4 h-4" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-              <path strokeLinecap="round" strokeLinejoin="round" d="M6 18L18 6M6 6l12 12" />
-            </svg>
-          </button>
-        </div>
-
-        {/* Body */}
-        <form onSubmit={handleSubmit} className="px-5 pb-5">
-          <div className="mb-4">
-            <label htmlFor="link-url" className="block text-sm font-medium text-slate-600 dark:text-slate-400 mb-1.5">
-              URL
-            </label>
-            <input
-              ref={inputRef}
-              id="link-url"
-              type="url"
-              value={url}
-              onChange={(e) => setUrl(e.target.value)}
-              placeholder="https://example.com"
-              className="input"
-            />
-          </div>
-
-          <div className="flex gap-2.5">
-            {initialUrl && (
-              <button
-                type="button"
-                onClick={handleRemoveLink}
-                className="btn-secondary px-3 py-2 text-sm text-rose-600 dark:text-rose-400"
-              >
-                Remove
-              </button>
-            )}
-            <div className="flex-1" />
-            <button
-              type="button"
-              onClick={onClose}
-              className="btn-secondary px-4 py-2 text-sm"
-            >
-              Cancel
-            </button>
-            <button
-              type="submit"
-              className="btn-primary px-4 py-2 text-sm"
-            >
-              {initialUrl ? 'Update' : 'Add Link'}
-            </button>
-          </div>
-        </form>
-      </div>
-    </div>
-  );
-}
-
-// Toolbar icons (prefixed TB to avoid collision with FloatingToolbar icons)
-function TBBoldIcon() {
-  return (
-    <svg className="w-4 h-4" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2.5}>
-      <path strokeLinecap="round" strokeLinejoin="round" d="M6 4h8a4 4 0 014 4 4 4 0 01-4 4H6z" />
-      <path strokeLinecap="round" strokeLinejoin="round" d="M6 12h9a4 4 0 014 4 4 4 0 01-4 4H6z" />
-    </svg>
-  );
-}
-
-function TBItalicIcon() {
-  return (
-    <svg className="w-4 h-4" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-      <path strokeLinecap="round" strokeLinejoin="round" d="M10 4h4m-2 0l-4 16m0 0h4" />
-    </svg>
-  );
-}
-
-function TBUnderlineIcon() {
-  return (
-    <svg className="w-4 h-4" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-      <path strokeLinecap="round" strokeLinejoin="round" d="M7 4v7a5 5 0 0010 0V4M5 21h14" />
-    </svg>
-  );
-}
-
-function TBStrikeIcon() {
-  return (
-    <svg className="w-4 h-4" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-      <path strokeLinecap="round" strokeLinejoin="round" d="M17 12H7m10-4a4 4 0 00-4-4H9a4 4 0 000 8m2 8a4 4 0 004-4" />
-    </svg>
-  );
-}
-
-function TBBulletListIcon() {
-  return (
-    <svg className="w-4 h-4" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-      <path strokeLinecap="round" strokeLinejoin="round" d="M8.25 6.75h9m-9 5.25h9m-9 5.25h9M4.5 6.75h.007v.008H4.5V6.75zm.375 0a.375.375 0 11-.75 0 .375.375 0 01.75 0zM4.5 12h.007v.008H4.5V12zm.375 0a.375.375 0 11-.75 0 .375.375 0 01.75 0zm-.375 5.25h.007v.008H4.5v-.008zm.375 0a.375.375 0 11-.75 0 .375.375 0 01.75 0z" />
-    </svg>
-  );
-}
-
-function TBOrderedListIcon() {
-  return (
-    <svg className="w-4 h-4" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-      <path strokeLinecap="round" strokeLinejoin="round" d="M8.25 6.75h9m-9 5.25h9m-9 5.25h9" />
-      <text x="3" y="8.5" fontSize="6" fill="currentColor" fontFamily="sans-serif">1</text>
-      <text x="3" y="14" fontSize="6" fill="currentColor" fontFamily="sans-serif">2</text>
-      <text x="3" y="19.5" fontSize="6" fill="currentColor" fontFamily="sans-serif">3</text>
-    </svg>
-  );
-}
-
-function TBLinkIcon() {
-  return (
-    <svg className="w-4 h-4" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-      <path strokeLinecap="round" strokeLinejoin="round" d="M13.828 10.172a4 4 0 00-5.656 0l-4 4a4 4 0 105.656 5.656l1.102-1.101m-.758-4.899a4 4 0 005.656 0l4-4a4 4 0 00-5.656-5.656l-1.1 1.1" />
-    </svg>
-  );
-}
-
-function TBEmojiIcon() {
-  return (
-    <svg className="w-4 h-4" fill="none" viewBox="0 0 24 24" stroke="currentColor" strokeWidth={2}>
-      <path strokeLinecap="round" strokeLinejoin="round" d="M14.828 14.828a4 4 0 01-5.656 0M9 10h.01M15 10h.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z" />
-    </svg>
-  );
-}
-
-function TBHeading2Icon() {
-  return (
-    <svg className="w-4 h-4" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth={2}>
-      <path strokeLinecap="round" strokeLinejoin="round" d="M4 6v12M4 12h7M11 6v12" />
-      <text x="15" y="19" fontSize="10" fill="currentColor" stroke="none" fontWeight="bold" fontFamily="sans-serif">2</text>
-    </svg>
-  );
-}
-
-function TBHeading3Icon() {
-  return (
-    <svg className="w-4 h-4" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth={2}>
-      <path strokeLinecap="round" strokeLinejoin="round" d="M4 6v12M4 12h7M11 6v12" />
-      <text x="15" y="19" fontSize="10" fill="currentColor" stroke="none" fontWeight="bold" fontFamily="sans-serif">3</text>
-    </svg>
-  );
-}
-
-function TBBlockquoteIcon() {
-  return (
-    <svg className="w-4 h-4" viewBox="0 0 24 24" fill="currentColor" stroke="none">
-      <path d="M6 17h3l2-4V7H5v6h3l-2 4zm8 0h3l2-4V7h-6v6h3l-2 4z" />
-    </svg>
-  );
-}
-
-function TBTaskListIcon() {
-  return (
-    <svg className="w-4 h-4" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth={2}>
-      <rect x="3" y="5" width="4" height="4" rx="0.5" strokeLinecap="round" strokeLinejoin="round" />
-      <path strokeLinecap="round" strokeLinejoin="round" d="M4.25 7l0.75 0.75L6.5 6" />
-      <path strokeLinecap="round" strokeLinejoin="round" d="M10 7h10" />
-      <rect x="3" y="14" width="4" height="4" rx="0.5" strokeLinecap="round" strokeLinejoin="round" />
-      <path strokeLinecap="round" strokeLinejoin="round" d="M10 16h10" />
-    </svg>
   );
 }
